finished the first assignment. all set

+++ Javascript/w1/index.php
@@ -30,23 +30,36 @@
       </div>
       <!--add aqui a tabela-->
       <div>
-        <table class="tableshow">
-          <tr>
-            <th>Name</th>
-            <th>Headline</th>
-            <th>Action</th>
-          </tr>
           <?php
             $stmt = $pdo->query("SELECT first_name, last_name, headline, profile_id FROM Profile");
-            while($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
-              echo "<tr><td>";
-              echo("<a href=view.php?profile_id=".$row['profile_id'].">".
-              (htmlentities($row['first_name'].' '.$row['last_name']))."</a>");
-              echo "</td><td>";
-              echo(htmlentities($row['headline']));
-              echo "</td><td>";
-              echo('<a href="edit.php?profile_id='.$row['profile_id'].'">Edit</a>'." / ".'<a href="delete.php?profile_id='.$row['profile_id'].'">Delete</a>');
-              echo "</td></tr>\n";
+            $row = $stmt->fetch(PDO::FETCH_ASSOC);
+            if ($row !== false){
+              ?>
+              <table class="tableshow">
+                <tr>
+                  <th>Name</th>
+                  <th>Headline</th>
+                  <th>Action</th>
+                </tr>
+                <?php
+                echo "<tr><td>";
+                echo("<a href=view.php?profile_id=".$row['profile_id'].">".
+                (htmlentities($row['first_name'].' '.$row['last_name']))."</a>");
+                echo "</td><td>";
+                echo(htmlentities($row['headline']));
+                echo "</td><td>";
+                echo('<a href="edit.php?profile_id='.$row['profile_id'].'">Edit</a>'." / ".'<a href="delete.php?profile_id='.$row['profile_id'].'">Delete</a>');
+                echo "</td></tr>\n";
+              while($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+                echo "<tr><td>";
+                echo("<a href=view.php?profile_id=".$row['profile_id'].">".
+                (htmlentities($row['first_name'].' '.$row['last_name']))."</a>");
+                echo "</td><td>";
+                echo(htmlentities($row['headline']));
+                echo "</td><td>";
+                echo('<a href="edit.php?profile_id='.$row['profile_id'].'">Edit</a>'." / ".'<a href="delete.php?profile_id='.$row['profile_id'].'">Delete</a>');
+                echo "</td></tr>\n";
+              }
             }
           ?>
         </table>
@@ -61,21 +74,32 @@
       <a href=login.php>Please log in</a>
     </div>
     <div>
-      <table class="tableshow">
-        <tr>
-          <th>Name</th>
-          <th>Headline</th>
-        </tr>
       <?php
         $stmt = $pdo->query("SELECT first_name, last_name, headline, profile_id FROM Profile");
-        while($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
-          echo "<tr><td>";
-          echo("<a href=view.php?profile_id=".$row['profile_id'].">".
-          (htmlentities($row['first_name'].' '.$row['last_name']))."</a>");
-          echo "</td><td>";
-          echo(htmlentities($row['headline']));
-          echo "</td><tr>\n";
-        }
+        $row = $stmt->fetch(PDO::FETCH_ASSOC);
+          if($row !== false) {
+            ?>
+            <table class="tableshow">
+              <tr>
+                <th>Name</th>
+                <th>Headline</th>
+              </tr>
+            <?php
+            echo "<tr><td>";
+            echo("<a href=view.php?profile_id=".$row['profile_id'].">".
+            (htmlentities($row['first_name'].' '.$row['last_name']))."</a>");
+            echo "</td><td>";
+            echo(htmlentities($row['headline']));
+            echo "</td><tr>\n";
+            while($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+              echo "<tr><td>";
+              echo("<a href=view.php?profile_id=".$row['profile_id'].">".
+              (htmlentities($row['first_name'].' '.$row['last_name']))."</a>");
+              echo "</td><td>";
+              echo(htmlentities($row['headline']));
+              echo "</td><tr>\n";
+            }
+          }
       ?>
     </table>
     </div>
